apps/api/pyproject.toml
[project]
name = "ecoalarm-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.8.0",
    "sqlalchemy>=2.0.0",
    "pyjwt>=2.9.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "httpx>=0.27.0"]
